move bert to synapses; move examples to examples. add network_dim item
