+++ bittensor/synapses/mnist/Dockerfile
